# frontend/openapi.snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "高考志愿填报智能体",
    "description": "基于位次法的可解释、可回测志愿推荐系统（数据层当前为模拟数据，严禁真实填报）。所有响应含 data / evidence / warnings 三段。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/students": {
      "post": {
        "tags": [
          "students"
        ],
        "summary": "创建档案",
        "operationId": "create_student_api_v1_students_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_StudentPayload_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/students/{student_id}": {
      "get": {
        "tags": [
          "students"
        ],
        "summary": "读取档案",
        "operationId": "get_student_api_v1_students__student_id__get",
        "parameters": [
          {
            "name": "student_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Student Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_StudentPayload_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "students"
        ],
        "summary": "补全/修改档案",
        "operationId": "patch_student_api_v1_students__student_id__patch",
        "parameters": [
          {
            "name": "student_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Student Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentPatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_StudentPayload_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/students/{student_id}/resolve-rank": {
      "post": {
        "tags": [
          "students"
        ],
        "summary": "分数 → 位次",
        "operationId": "resolve_rank_api_v1_students__student_id__resolve_rank_post",
        "parameters": [
          {
            "name": "student_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Student Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_ResolveRankPayload_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/plans/generate": {
      "post": {
        "tags": [
          "plans"
        ],
        "summary": "生成志愿表",
        "operationId": "generate_plan_api_v1_plans_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_PlanPayload_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/plans/{plan_id}": {
      "get": {
        "tags": [
          "plans"
        ],
        "summary": "读取志愿表",
        "operationId": "get_plan_api_v1_plans__plan_id__get",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_PlanPayload_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/plans/{plan_id}/items": {
      "patch": {
        "tags": [
          "plans"
        ],
        "summary": "手改志愿表（覆盖式）",
        "operationId": "patch_items_api_v1_plans__plan_id__items_patch",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanPatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_PlanPayload_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/plans/{plan_id}/validate": {
      "post": {
        "tags": [
          "plans"
        ],
        "summary": "风险扫描（重跑）",
        "operationId": "validate_plan_api_v1_plans__plan_id__validate_post",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_PlanPayload_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/plans/{plan_id}/export": {
      "get": {
        "tags": [
          "plans"
        ],
        "summary": "导出报告（pdf | xlsx）",
        "operationId": "export_plan_api_v1_plans__plan_id__export_get",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(pdf|xlsx)$",
              "description": "pdf 或 xlsx",
              "default": "pdf",
              "title": "Format"
            },
            "description": "pdf 或 xlsx"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/risk/scan": {
      "post": {
        "tags": [
          "risk"
        ],
        "summary": "志愿风险速查",
        "operationId": "scan_risks_api_v1_risk_scan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RiskScanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_RiskScanPayload_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Envelope_PlanPayload_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PlanPayload"
          },
          "evidence": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Evidence"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "Envelope[PlanPayload]"
      },
...
      "Envelope_ResolveRankPayload_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ResolveRankPayload"
          },
          "evidence": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Evidence"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "Envelope[ResolveRankPayload]"
      },
      "Envelope_RiskScanPayload_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/RiskScanPayload"
          },
          "evidence": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Evidence"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "Envelope[RiskScanPayload]"
      },
      "Envelope_StudentPayload_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/StudentPayload"
          },
          "evidence": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Evidence"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "Envelope[StudentPayload]"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PlanGenerateRequest": {
        "properties": {
          "student_id": {
            "type": "string",
            "title": "Student Id"
          },
          "filters": {
            "$ref": "#/components/schemas/RecommendFilters"
          },
          "weights": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weights"
          },
          "plan_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Id"
          },
          "preference_order": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preference Order",
            "description": "考生意愿序（unit_id 列表）；不填则按冲稳保垫 + 层内效用排序"
          },
          "obey_adjustment": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Obey Adjustment",
            "description": "院校专业组模式是否服从调剂；专业+院校模式忽略"
          }
        },
        "type": "object",
        "required": [
          "student_id"
        ],
        "title": "PlanGenerateRequest"
      },
...
      "PlanItemPatch": {
        "properties": {
          "unit_id": {
            "type": "string",
            "title": "Unit Id"
          },
          "obey_adjustment": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Obey Adjustment"
          }
        },
        "type": "object",
        "required": [
          "unit_id"
        ],
        "title": "PlanItemPatch"
      },
      "PlanPatchRequest": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PlanItemPatch"
            },
            "type": "array",
            "title": "Items",
            "description": "完整有序志愿列表（覆盖式）"
          },
          "obey_adjustment": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Obey Adjustment",
            "description": "统一设置服从调剂（覆盖逐项值）"
          },
          "filters": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RecommendFilters"
              },
              {
                "type": "null"
              }
            ],
            "description": "生成该志愿表时使用的筛选条件。手改只允许在**同口径重新评估出的候选池**内增删，带上它才能保证「池」与你当初看到的推荐列表完全一致（不带则视为无筛选）。"
          }
        },
        "type": "object",
        "title": "PlanPatchRequest"
      },
...
      "RiskScanRequest": {
        "properties": {
          "student_id": {
            "type": "string",
            "title": "Student Id"
          },
          "unit_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Unit Ids"
          },
          "obey_adjustment": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Obey Adjustment"
          },
          "obey_adjustment_map": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "boolean"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Obey Adjustment Map"
          }
        },
        "type": "object",
        "required": [
          "student_id",
          "unit_ids"
        ],
        "title": "RiskScanRequest"
      },
...
      "StudentCreateRequest": {
        "properties": {
          "province": {
            "type": "string",
            "title": "Province"
          },
          "year": {
            "type": "integer",
            "title": "Year",
            "default": 2026
          },
          "track": {
            "type": "string",
            "title": "Track",
            "default": "综合"
          },
          "subjects": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Subjects"
          },
          "total_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Score"
          },
          "rank": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rank"
          },
          "gender": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gender"
          },
          "is_fresh_graduate": {
            "type": "boolean",
            "title": "Is Fresh Graduate",
            "default": true
          },
          "political_status": {
            "type": "string",
            "title": "Political Status",
            "default": "群众"
          },
          "foreign_language": {
            "type": "string",
            "title": "Foreign Language",
            "default": "英语"
          },
          "single_subject_scores": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Single Subject Scores"
          },
          "physical_exam": {
            "$ref": "#/components/schemas/PhysicalExam"
          },
          "bonus_points": {
            "type": "integer",
            "title": "Bonus Points",
            "default": 0
          },
          "bonus_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bonus Type"
          },
          "preferences": {
            "$ref": "#/components/schemas/Preferences"
          }
        },
        "type": "object",
        "required": [
          "province"
        ],
        "title": "StudentCreateRequest",
        "description": "建档（允许草稿态：缺的字段会在响应 ``missing_fields`` 里给出）。"
      },
      "StudentPatchRequest": {
        "properties": {
          "province": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Province"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          },
          "track": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Track"
          },
          "subjects": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subjects"
          },
          "total_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Score"
          },
          "rank": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rank"
          },
          "gender": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gender"
          },
          "is_fresh_graduate": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Fresh Graduate"
          },
          "political_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Political Status"
          },
          "foreign_language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Foreign Language"
          },
          "single_subject_scores": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "integer"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Single Subject Scores"
          },
          "physical_exam": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PhysicalExam"
              },
              {
                "type": "null"
              }
            ]
          },
          "bonus_points": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bonus Points"
          },
          "bonus_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bonus Type"
          },
          "preferences": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Preferences"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "StudentPatchRequest",
        "description": "增量补全（只提交要改的字段；未提交字段保持原值）。"
      },
...
    }
  }
}